[Rename] Suggest View Type 수정 및 API 함수 재정리
