+++ API/favorit/deleteFavorit.ts
@@ -1,5 +1,5 @@
 import { FavoritProps } from "./addFavorit";
-import { baseInstance } from "./instance";
+import { baseInstance } from "../instance";
 
 export const deleteFavorit = async (props: FavoritProps) => {
   const request = {userId: props.userId}
