+++ components/StoreDetailView/ReviewBox.tsx
@@ -2,10 +2,10 @@ import { StyleSheet, View, Dimensions, Alert } from "react-native";
 import UserIcon from "react-native-vector-icons/FontAwesome";
 import CustomText from "../CustomText";
 import Carousel from "./Carousel";
-import { type ReviewType } from "../../API/getReviewList";
+import { type ReviewType } from "../../API/review/getReviewList";
 import DeleteIcon from "react-native-vector-icons/Feather";
 import EditIcon from "react-native-vector-icons/AntDesign";
-import deleteReview from "../../API/deleteReview";
+import deleteReview from "../../API/review/deleteReview";
 import { NavigationProp } from "@react-navigation/native";
 import { RootStackParamList } from "../../App";
 import RatingStar from "../RatingStar";
